+++ app/src/main/java/woowacourse/shopping/data/product/ProductEntity.kt
@@ -3,7 +3,6 @@ package woowacourse.shopping.data.product
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import woowacourse.shopping.domain.product.Product
 
 @Entity(tableName = "cart")
 data class ProductEntity(
@@ -12,11 +11,3 @@ data class ProductEntity(
     @ColumnInfo("name") val name: String,
     @ColumnInfo("price") val price: Int,
 )
-
-fun Product.toProductEntity(): ProductEntity {
-    return ProductEntity(
-        imageUrl = this.imageUrl,
-        name = this.name,
-        price = this.price,
-    )
-}
